requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt
mysql-connector-python
requests
python-dotenv
gunicorn
PyPDF2
pdfplumber
pdf2image
pytesseract
Pillow
werkzeug
